# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Asistente Personal Modular API
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /automation/rules:
    post:
      summary: Create automation rule
      responses:
        "201": { description: Created }
        "412": { description: Approval required for sensitive actions }
    get:
      summary: List rules
      responses:
        "200": { description: OK }
  /automation/rules/{id}/test:
    post:
      summary: Test rule execution
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Run log }
...
  /policy/approvals/{id}/approve:
    post:
      summary: Approve sensitive operation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Approved }
  /policy/approvals/{id}/reject:
    post:
      summary: Reject sensitive operation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Rejected }
...